refactoring overview related code
