+++ freeplane/src/editor/resources/translations/Resources_zh_CN.properties
@@ -829,8 +829,7 @@ map_load_error=\u4E0D\u80FD\u8F7D\u5165\u601D\u7EF4\u5BFC\u56FE {0}
 map_locked_by_open=\u601D\u7EF4\u5BFC\u56FE {0} \u6B63\u88AB\u7528\u6237 {1} \u7F16\u8F91, \u5C06\u4EE5\u53EA\u8BFB\u65B9\u5F0F\u6253\u5F00.
 map_locked_by_save_as=\u601D\u7EF4\u5BFC\u56FE {0} \u6B63\u88AB\u7528\u6237 {1} \u7F16\u8F91, \u4FDD\u5B58\u52A8\u4F5C\u88AB\u4E2D\u6B62.
 map_not_saved=\u8BE5\u5BFC\u56FE\u4ECE\u672A\u4FDD\u5B58\u8FC7.
-map_overview=\u7F29\u7565\u56FE
-map_overview_attach_point=\u8BBE\u7F6E\u505C\u9760\u70B9
+map_overview_attach_point=\u7F29\u7565\u56FE\u505C\u9760\u70B9
 MapBackgroundClearAction.text=\u6E05\u9664
 MapBackgroundColorAction.text=\u80CC\u666F\u989C\u8272
 MapBackgroundImageAction.text=\u80CC\u666F\u56FE\u7247\u2026
@@ -1251,7 +1250,6 @@ OptionPanel.map_overview_attach_point.NORTH_EAST=\u53F3\u4E0A
 OptionPanel.map_overview_attach_point.NORTH_WEST=\u5DE6\u4E0A
 OptionPanel.map_overview_attach_point.SOUTH_EAST=\u53F3\u4E0B
 OptionPanel.map_overview_attach_point.SOUTH_WEST=\u5DE6\u4E0B
-OptionPanel.map_overview_bounds.=\u91CD\u7F6E\u5927\u5C0F\u548C\u539F\u70B9
 OptionPanel.markClones=\u4E3A\u514B\u9686\u6807\u8BB0\u4E0A\u8272
 OptionPanel.markdown_disable_editor=\u7981\u7528Markdown\u7F16\u8F91\u5668
 OptionPanel.markdown_editor_font=Markdown\u7F16\u8F91\u5668\u5B57\u4F53
