+++ freeplane/src/editor/resources/translations/Resources_zh_TW.properties
@@ -821,8 +821,7 @@ map_load_error=\u7121\u6CD5\u7121\u6CD5\u8F09\u5165\u5FC3\u667A\u5716 {0}
 map_locked_by_open=\u5FC3\u667A\u5716 {0} \u5DF2\u7D93\u88AB\u4F7F\u7528\u8005 {1} \u7DE8\u8F2F\uFF0C\u5C07\u4EE5\u552F\u8B80\u7684\u72C0\u614B\u958B\u555F\u3002
 map_locked_by_save_as=\u5FC3\u667A\u5716 {0} \u6B63\u88AB\u4F7F\u7528\u8005 {1} \u7DE8\u8F2F\uFF0C\u4E2D\u6B62\u53E6\u5B58\u65B0\u6A94\u3002
 map_not_saved=\u8A72\u5FC3\u667A\u5716\u5C1A\u672A\u5B58\u6A94\u3002
-map_overview=\u7E2E\u7565\u5716
-map_overview_attach_point=\u8A2D\u7F6E\u505C\u9760\u9EDE
+map_overview_attach_point=\u7E2E\u7565\u5716\u505C\u9760\u9EDE
 MapBackgroundClearAction.text=\u660E\u4EAE
 MapBackgroundColorAction.text=\u80CC\u666F\u984F\u8272
 MapBackgroundImageAction.text=\u80CC\u666F\u5716\u7247\u2026
@@ -1219,7 +1218,6 @@ OptionPanel.map_overview_attach_point.NORTH_EAST=\u53F3\u4E0A
 OptionPanel.map_overview_attach_point.NORTH_WEST=\u5DE6\u4E0A
 OptionPanel.map_overview_attach_point.SOUTH_EAST=\u53F3\u4E0B
 OptionPanel.map_overview_attach_point.SOUTH_WEST=\u5DE6\u4E0B
-OptionPanel.map_overview_bounds.=\u91CD\u7F6E\u5927\u5C0F\u548C\u539F\u9EDE
 OptionPanel.markClones=\u7E6A\u88FD\u8907\u88FD\u6A19\u8A18
 OptionPanel.markSelectedClones=\u70BA\u9078\u64C7\u7E6A\u88FD\u8907\u88FD\u6A19\u8A18
 OptionPanel.max_displayed_node_count=\u6700\u5927\u986F\u793A\u7BC0\u9EDE\u6578
