+++ freeplane/src/main/java/org/freeplane/view/swing/map/overview/MapOverviewAttachPoint.java
@@ -1,0 +1,5 @@
+package org.freeplane.view.swing.map.overview;
+
+enum MapOverviewAttachPoint {
+    SOUTH_EAST, SOUTH_WEST, NORTH_EAST, NORTH_WEST
+}
