+++ freeplane/src/viewer/resources/translations/Resources_en.properties
@@ -829,8 +829,7 @@ map_load_error=Could not load map {0}
 map_locked_by_open=The map {0} is already being edited by the user {1}. Opening as read-only.
 map_locked_by_save_as=The map {0} is being edited by the user {1}. Action Save As aborted.
 map_not_saved=The map was not saved before.
-map_overview=Map overview
-map_overview_attach_point=Set attach point
+map_overview_attach_point=Map overview attach point
 MapBackgroundClearAction.text=Clear
 MapBackgroundColorAction.text=Background color
 MapBackgroundImageAction.text=Background Image\u2026
@@ -1258,7 +1257,6 @@ OptionPanel.map_overview_attach_point.NORTH_EAST=Top right
 OptionPanel.map_overview_attach_point.NORTH_WEST=Top left
 OptionPanel.map_overview_attach_point.SOUTH_EAST=Bottom right
 OptionPanel.map_overview_attach_point.SOUTH_WEST=Bottom left
-OptionPanel.map_overview_bounds.=Reset size and origin
 OptionPanel.markClones=Paint clone marks
 OptionPanel.markdown_disable_editor=Disable Markdown Editor
 OptionPanel.markdown_editor_font=Font for Markdown Editor
